requirements.txt
langchain==0.2.11
langchain-groq==0.1.7
langchain-community==0.2.10
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
pypdf==4.3.1
sentence-transformers==3.0.1
streamlit==1.37.0
streamlit-chat==0.1.1
python-dotenv==1.0.1
scikit-learn==1.5.1
rank_bm25==0.2.2
numpy==1.26.4